Add DualActionButton widget for seek + bookmark buttons
This widget allows to have an easy-to-use API for having
a button that responds to a "default" action and another
action that is executed when the button is pressed a bit
longer ("longpress").

The widget's image/label will be set to an alternative
value, so it "shows" to the user which action will be
carried out.

The longpress action of the bookmark button is to switch
to the playlist tab and focus the currently-playing track.