Make panucci.py more modular
This is going to break things.

GTK_Main has been broken up into PanucciGUI and PlayerTab which makes the
code (imo) much easier to understand. What's really cool is that both
PlayerTab and PlaylistTab are now completely optional, Panucci will run
without them \o/  (not quite sure what this will be useful for, but it
demonstrates good modularity)

NOTE: This is a _very_ rough patch, but I'm releasing early so people can
      get a look at it and more importantly so I don't procrasinate and leave
      this mess sitting in the repo for too long.